backend/requirement.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
copilotkit
ag-ui-langgraph
python-dotenv